fix(scripts): scrub persisted inline run output; simplify script-runs UI
Addresses the Codex P1 review on #666 plus cleanup-review findings.

- P1 (security): scrub inline run `args` + `output.result` before persisting
  to `script_runs`. The row is later served raw by GET /api/script-runs/{id}
  to the dashboard, so it needs the same redaction guarantee the run response
  already has (the immediate response + error were scrubbed; the stored copy
  was not).
- Consolidate the forked `formatDurationMs`: one implementation in
  `@/lib/format-duration-ms` with an opt-in `precise` mode (each mode is
  byte-identical to its prior source); `step-shared` now delegates instead of
  re-defining a same-named function with divergent rounding.
- Hoist the duplicated newline-index + binary-search `lineOf` in `source-map`
  into a shared `lineIndexer(source)` used by both parsers.
- Replace hand-rolled clipboard copy state in `RunId` and `SourceView` with
  the existing `useCopyToClipboard` hook.
